Rename specs to specifications and add explicit imports
