+++ examples/devices/amplifier.py
@@ -5,7 +5,7 @@ from typing_extensions import TypedDict
 from tickit.adapters.epics import EpicsAdapter
 from tickit.adapters.io.epics_io import EpicsIo
 from tickit.adapters.io.tcp_io import TcpIo
-from tickit.adapters.specs.regex_command import RegexCommand
+from tickit.adapters.specifications import RegexCommand
 from tickit.adapters.tcp import CommandAdapter
 from tickit.core.adapter import AdapterContainer
 from tickit.core.components.component import Component, ComponentConfig
