+++ examples/devices/remote_controlled.py
@@ -6,7 +6,7 @@ from typing import AsyncIterable, Optional, TypedDict
 import pydantic.v1.dataclasses
 
 from tickit.adapters.io.tcp_io import TcpIo
-from tickit.adapters.specs.regex_command import RegexCommand
+from tickit.adapters.specifications import RegexCommand
 from tickit.adapters.tcp import CommandAdapter
 from tickit.core.adapter import AdapterContainer
 from tickit.core.components.component import Component, ComponentConfig
