+++ src/tickit/adapters/specifications/__init__.py
@@ -1,0 +1,4 @@
+from .http_endpoint import HttpEndpoint
+from .regex_command import RegexCommand
+
+__all__ = ["HttpEndpoint", "RegexCommand"]
