+++ src/tickit/adapters/specifications/http_endpoint.py
