+++ src/tickit/adapters/specifications/regex_command.py
